Updated query handling
