+++ src/utils/query/createTermFilter.ts
@@ -1,0 +1,11 @@
+export const createTermFilter = (field: string, value: string) => {
+  return {
+    op: 'in',
+    content: {
+      field,
+      value: [value],
+    },
+  };
+};
+
+export default createTermFilter;
